+++ src/components/LayoutDefault.vue
@@ -1,19 +1,28 @@
 <template>
-  <div>
-    <app-sidebar />
+  <div class="layout-default">
+    <navigation-sidebar />
+    <main class="main-content">
+      <slot />
+    </main>
   </div>
 </template>
 
 <script>
-import { defineComponent } from "vue"
-import AppSidebar from "@/components/AppSidebar"
+import { defineComponent } from "vue";
+import NavigationSidebar from "@/components/NavigationSidebar";
 
 export default defineComponent({
   name: "LayoutDefault",
-  components: { AppSidebar }
-})
+  components: { NavigationSidebar }
+});
 </script>
 
 <style scoped>
+.layout-default {
+  display: flex;
+}
 
+.main-content {
+  margin: 1rem;
+}
 </style>