+++ src/pages/Users.vue
@@ -1,17 +1,18 @@
 <template>
   <div>
+    <h1>Users page</h1>
     <list-default />
   </div>
 </template>
 
 <script lang="ts">
-import { defineComponent } from "vue";
-import ListDefault from "@/components/ListDefault.vue"
+import { defineComponent } from 'vue';
+import ListDefault from '@/components/ListDefault.vue';
 
 export default defineComponent({
-  name: "UsersMainPage",
+  name: 'UsersPage',
   components: { ListDefault }
-})
+});
 </script>
 
 <style scoped>