+++ src/router/users.ts
@@ -1,9 +1,14 @@
-const UsersMainPage = () => import('@/pages/UsersMain.vue')
+import { RouteName } from '@/specification/routes';
+import { routes } from '@/specification/routes';
+
+const usersPageName: RouteName = 'users';
+
+const UsersPage = () => import('@/pages/Users.vue');
 
 export default [
   {
-    path: '/users',
-    name: 'users',
-    component: UsersMainPage
+    path: routes[usersPageName],
+    name: usersPageName,
+    component: UsersPage
   }
-]
+];