+++ src/specification/routes.ts
@@ -1,1 +1,8 @@
-export type Routes = 'mainPage' | 'users' | 'comments' | 'photos'
+export const routes = {
+    mainPage: '/',
+    photos: '/photos',
+    comments: '/comments',
+    users: '/users'
+  } as const;
+
+export type RouteName = keyof typeof routes;