+++ src/store/index.ts
@@ -4,12 +4,12 @@ export const store = createStore({
   state () {
     return {
       count: 0
-    }
+    };
   },
   mutations: {
     increment (state) {
       // @ts-ignore
-      state.count++
+      state.count++;
     }
   }
-})
+});